+++ code/midtown/agi/texdef.cpp
@@ -83,9 +83,9 @@ void agiTexDef::CheckSurface()
 }
 
 // Same as DataCache alignment
-static inline constexpr usize AlignSize(usize value) noexcept
+static inline constexpr u32 AlignSize(u32 value) noexcept
 {
-    return (value + 7) & ~usize(7); // FIXME: 64-bit requires 16-byte alignment
+    return (value + 7) & ~u32(7); // FIXME: 64-bit requires 16-byte alignment
 }
 
 void agiTexDef::DoPageIn()
