+++ code/midtown/agiworld/meshsave.cpp
@@ -28,9 +28,9 @@ define_dummy_symbol(agiworld_meshsave);
 #include "vector7/vector4.h"
 
 // Same as DataCache alignment
-static inline constexpr usize AlignSize(usize value) noexcept
+static inline constexpr u32 AlignSize(u32 value) noexcept
 {
-    return (value + 7) & ~usize(7); // FIXME: 64-bit requires 16-byte alignment
+    return (value + 7) & ~u32(7); // FIXME: 64-bit requires 16-byte alignment
 }
 
 void agiMeshSet::BinarySave(Stream* stream)
